+++ aim/web/ui/src/components/ImagesPanel/ImagesPanel.d.ts
@@ -2,7 +2,7 @@ import React from 'react';
 
 export interface IImagesPanelProps {
   imagesData: any;
-  orderingData: { [key: string]: any };
+  orderedMap: { [key: string]: any };
   imagesBlobs: { [key: string]: value };
   recordSlice: number[];
   indexSlice: number[];